% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{as_file}
\alias{as_file}
\title{Mark as a file by adding classes}
\usage{
as_file(path)
}
\arguments{
\item{path}{file path}
}
\value{
\code{path} with classes set to \code{"file"} and \code{"file_{ext}"} where \code{ext} is
the file extension
}
\description{
Mark as a file by adding classes
}
\examples{
tf <- tempfile(fileext = ".txt")
writeLines("", tf)

as_file(tf)

}
