% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_is.R
\name{as_is}
\alias{as_is}
\title{Mark as a node to inject as is}
\usage{
as_is(x, error_call = current_env())
}
\arguments{
\item{x}{raw ooxml passed through \code{\link[=as_xml_node]{as_xml_node()}}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
adds the class "as_is" to the input text the content "as is" and not
modify it when polishing
}
\description{
Mark as a node to inject as is
}
