% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_is.R
\name{as_md}
\alias{as_md}
\title{Mark text as markdown}
\usage{
as_md(x)
}
\arguments{
\item{x}{text to be interpreted as markdown}
}
\value{
adds the class "as_ms" to the input text the content "as markdown"
apply markdown formatting to the text when polishing
}
\description{
Mark text as markdown
}
