% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{as_xml_nodeset}
\alias{as_xml_nodeset}
\alias{as_xml_node}
\alias{as_xml_node_word}
\alias{as_xml_node_pptx}
\alias{as_xml_word}
\alias{as_xml_pptx}
\title{Convert to xml nodes}
\usage{
as_xml_nodeset(
  x,
  ns = c("word", "pptx"),
  ...,
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  error_call = current_env()
)

as_xml_node(
  x,
  ...,
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  error_call = current_env()
)

as_xml_node_word(
  x,
  ...,
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  error_call = current_env()
)

as_xml_node_pptx(
  x,
  ...,
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  error_call = current_env()
)

as_xml_word(
  x,
  ...,
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  error_call = current_env()
)

as_xml_pptx(
  x,
  ...,
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  error_call = current_env()
)
}
\arguments{
\item{x}{an object to convert to xml nodes}

\item{ns}{word or pptx}

\item{...}{These dots are for future extensions and must be empty.}

\item{.envir, .open, .close}{forwarded to \code{\link[glue:glue]{glue::glue()}}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
an \code{xml_nodeset}
}
\description{
Convert to xml nodes
}
