% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_polish_app.R
\name{create_polish_app}
\alias{create_polish_app}
\title{Create a ready-to-use Polish App skeleton}
\usage{
create_polish_app(
  app.dirname,
  example.modules = FALSE,
  helpful.comments = FALSE
)
}
\arguments{
\item{app.dirname}{New directory that will be created and contain the shiny app files}

\item{example.modules}{If \code{TRUE}, two example modules will be created along with the app skeleton}

\item{helpful.comments}{If \code{TRUE}, comments will be included to help the user learn how Polish Apps work}
}
\description{
Creates needed files and directories so that you can quickly begin development of your Shiny App.
}
\examples{
# Create a new polish app skeleton with example modules and helpful comments in the app.R file.
\dontrun{
polish::create_polish_app(
  app.dirname = file.path('.', 'my_new_shiny_app'),
  example.modules = TRUE,
  helpful.comments = TRUE
)
}
}
