% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularServer.R
\name{modularServer}
\alias{modularServer}
\title{Load server code into Shiny App from files}
\usage{
modularServer(module.dir, module.file, environment)
}
\arguments{
\item{module.dir}{Path to the parent directory containing module directories}

\item{module.file}{Name of file within module directories containing server code}

\item{environment}{\code{environment()} to be passed to \code{source()} (must be provided when function is called)}
}
\description{
Loops through modules, loading server code into Shiny via \code{source()} if files exist.
}
