% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularTabPanels.R
\name{modularTabPanels}
\alias{modularTabPanels}
\title{Load UI code into Shiny App from files}
\usage{
modularTabPanels(module.dir, module.file, numbered = FALSE)
}
\arguments{
\item{module.dir}{Path to directory that contains modules (child directories)}

\item{module.file}{Name of file found within module directories that contains UI code}

\item{numbered}{If \code{TRUE}, directories will be ordered by numeric prefix (e.g., \verb{01_Index}, \verb{02_About})}
}
\value{
A list of Shiny \code{tabPanel()} objects
}
\description{
Maps over modules, converting Shiny UI code into \code{tabPanel()} objects.
}
