% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\docType{package}
\name{polish-package}
\alias{polish-package}
\title{Polish up your R content.}
\description{
Polish up your R content (RMarkdown and flexdashboards) 
using attractive templates for publishing to RStudio Connect.
}
\author{
Will Bishop \email{wcmbishop@gmail.com}
}
