% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polish_check_dots_empty}
\alias{polish_check_dots_empty}
\title{Check ... empty}
\usage{
polish_check_dots_empty(env = caller_env(), error = NULL, call = caller_env())
}
\arguments{
\item{env}{Environment in which to look for \code{...}.}

\item{error}{An optional error handler passed to \code{\link[rlang:try_fetch]{try_fetch()}}. Use
this e.g. to demote an error into a warning.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Wrapper around \code{\link[rlang:check_dots_empty]{rlang::check_dots_empty()}} that only does the check
if the \code{polish.dots_error} is set to TRUE
}
