% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R, R/ggplot2.R, R/word.R
\name{polish_content_word.flextable}
\alias{polish_content_word.flextable}
\alias{polish_content_word.ggplot}
\alias{polish_content_word}
\alias{polish_content_word.character}
\alias{polish_content_word.file_png}
\title{Polish content for Word}
\usage{
\method{polish_content_word}{flextable}(
  x,
  inline = TRUE,
  ...,
  align = NULL,
  split = NULL,
  topcaption = TRUE,
  keepnext = NULL,
  error_call = current_env()
)

\method{polish_content_word}{ggplot}(
  x,
  inline = FALSE,
  ...,
  height = 5,
  width = 6,
  units = "in",
  res = 300,
  error_call = current_env()
)

polish_content_word(x, inline = FALSE, ..., error_call = current_env())

\method{polish_content_word}{character}(
  x,
  inline = FALSE,
  ...,
  font_color = NULL,
  font_style = NULL,
  font_size = NULL,
  font_typeface = NULL,
  error_call = current_env()
)

\method{polish_content_word}{file_png}(
  x,
  inline = FALSE,
  ...,
  height = 5,
  width = 6,
  units = "in",
  error_call = current_env()
)
}
\arguments{
\item{x}{object to be polished}

\item{inline}{if \code{FALSE} the comments are wrapped in a \verb{<w:p>}}

\item{...}{used in some methods}

\item{align}{alignment of table and caption}

\item{split}{should table be allowed to split}

\item{topcaption}{should caption be above or below the table}

\item{keepnext}{should content be marked for keep next.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
defaults to 5 inches tall and 6 inches wide.}

\item{res}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{font_color}{color of text. Enter as a hex code.}

\item{font_style}{styling of text, vector including "bold","italic","underline".}

\item{font_size}{point size of text if different from document default.}

\item{font_typeface}{typeface of the text to be added if different from document default.}
}
\description{
Polish content for Word
}
