% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_html.R
\name{polish_html}
\alias{polish_html}
\title{Convert to a Polish HTML Document.}
\usage{
polish_html(toc = TRUE, toc_float = TRUE, code_folding = "hide",
  contact = NULL, source = NULL, type = "default", home = NULL,
  ...)
}
\arguments{
\item{toc}{TRUE to include a table of contents in the output}

\item{toc_float}{TRUE to float the table of contents to the left of the main document content.}

\item{code_folding}{Enable document readers to toggle the display of R code chunks. 
Specify "none" to display all code chunks (assuming they 
were knit with echo = TRUE). Specify "hide" to hide all R 
code chunks by default (users can show hidden code chunks 
either individually or document-wide). Specify "show" to 
show all R code chunks by default.}

\item{contact}{email address (as a string) to show as a point of contact}

\item{source}{URL link (as a string) to your source code (e.g in GitHub)}

\item{type}{css styling type: either "default" or "inverse" (for inverse coloring)}

\item{home}{URL link (as string) to provide a link to a "home" site}

\item{...}{optional named inputs passed into \code{\link[rmarkdown]{html_document}}.}
}
\value{
R Markdown output format from \pkg{rmarkdown} 
\code{\link[rmarkdown]{html_document}}.
}
\description{
Custom format rendering that converts from R Markdown to a Polish style HTML
document. This is an extension of the \pkg{rmarkdown} 
\code{\link[rmarkdown]{html_document}} function.
}
