% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml_nodeset_ns_spec}
\alias{xml_nodeset_ns_spec}
\title{xml namespaces}
\usage{
xml_nodeset_ns_spec(ns = c("word", "pptx"), error_call = caller_env())
}
\arguments{
\item{ns}{word or pptx}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
xml namespaces for word or pptx
}
\description{
xml namespaces
}
