% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{polish_content}
\alias{polish_content}
\title{Polish content}
\usage{
polish_content(x, type, ..., error_call = current_env(), error_class = NULL)
}
\arguments{
\item{x}{object to be polished into ooxml}

\item{type}{output type. Valid values are "word" and "pptx".}

\item{...}{additional arguments for methods}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{error_class}{Additional error class.}
}
\value{
an object of class "polish_xml_nodeset". Contains XML that is viable Word or PowerPoint OOXML for addition
to Word documents or PowerPoint presentations based on the "type".
}
\description{
Polish content
}
\details{
depending on \code{type} either the \code{\link[=polish_content_pptx]{polish_content_pptx()}} or \code{\link[=polish_content_word]{polish_content_word()}}
function is called.
}
