% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R, R/ggplot2.R, R/pptx.R
\name{polish_content_pptx.flextable}
\alias{polish_content_pptx.flextable}
\alias{polish_content_pptx.ggplot}
\alias{polish_content_pptx}
\alias{polish_content_pptx.character}
\alias{polish_content_pptx.list}
\alias{polish_content_pptx.file_png}
\title{Polish content for Powerpoint}
\usage{
\method{polish_content_pptx}{flextable}(
  x,
  ph = "<p:ph/>",
  ...,
  full_width = FALSE,
  page_size = NULL,
  table_style = NULL,
  pptx = NULL,
  error_call = current_env()
)

\method{polish_content_pptx}{ggplot}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  height = 5,
  width = 6,
  units = "in",
  res = 300,
  error_call = current_env()
)

polish_content_pptx(x, ph = "<p:ph/>", pptx, ..., error_call = current_env())

\method{polish_content_pptx}{character}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  escape = TRUE,
  collapse = FALSE,
  font_color = NULL,
  font_style = NULL,
  font_size = NULL,
  font_typeface = NULL,
  error_call = current_env()
)

\method{polish_content_pptx}{list}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  list_type = c("none", "unordered", "ordered"),
  error_call = current_env()
)

\method{polish_content_pptx}{file_png}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  height = 5,
  width = 6,
  guess_size = FALSE,
  units = "in",
  error_call = current_env()
)
}
\arguments{
\item{x}{object to be polished}

\item{ph}{placeholder}

\item{...}{These dots are for future extensions and must be empty.}

\item{full_width}{Logical, used to fit full size of given width}

\item{page_size}{Number of rows to accommodate in a slide}

\item{table_style}{Named list with two entries: style and options. style is
the name of table style to apply as a character, options a character vector
of options to apply from the table style, including: "Banded Columns",
"Banded Rows", "First Column", "Header Row", "Last Column", and "Total Row"}

\item{pptx}{The powerpoint presentation.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
defaults to 5 inches tall and 6 inches wide.}

\item{res}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{escape}{Should \code{x} be escaped with \code{\link[htmltools:htmlEscape]{htmltools::htmlEscape()}}}

\item{collapse}{If \code{TRUE} a \verb{<a:br w:type="textWrapping" w:clear="all">} element is added between the \verb{<a:r>} nodes}

\item{font_color}{color of text. Enter as a hex code.}

\item{font_style}{styling of text, vector including "bold","italic","underline".}

\item{font_size}{point size of text if different from document default.}

\item{font_typeface}{typeface of the text to be added if different from document default.}

\item{list_type}{How to print the list. Options are "none", "unordered" and
"ordered".
\itemize{
\item "none" (default) puts contents on new lines and indents lists accordingly.
\item "unordered" makes the content bulleted list with indents.
\item "ordered" makes a numbered list with intents.
}}

\item{guess_size}{see \link[officer:external_img]{officer::external_img}}
}
\description{
Polish content for Powerpoint
}
