% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R, R/ggplot2.R, R/pptx.R
\name{polish_content_pptx.flextable}
\alias{polish_content_pptx.flextable}
\alias{polish_content_pptx.ggplot}
\alias{polish_content_pptx}
\alias{polish_content_pptx.character}
\alias{polish_content_pptx.list}
\alias{polish_content_pptx.file_png}
\title{Polish content for Powerpoint}
\usage{
\method{polish_content_pptx}{flextable}(
  x,
  ph = "<p:ph/>",
  ...,
  full_width = FALSE,
  page_size = NULL,
  table_style = NULL,
  pptx = NULL,
  error_call = current_env()
)

\method{polish_content_pptx}{ggplot}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  height = NULL,
  width = NULL,
  res = 300,
  device = c("png", "svg", "jpeg"),
  scale = 1,
  error_call = current_env()
)

polish_content_pptx(x, ph = "<p:ph/>", pptx, ..., error_call = current_env())

\method{polish_content_pptx}{character}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  escape = TRUE,
  collapse = FALSE,
  font_color = NULL,
  font_style = NULL,
  font_size = NULL,
  font_typeface = NULL,
  error_call = current_env()
)

\method{polish_content_pptx}{list}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  list_type = c("none", "unordered", "ordered"),
  error_call = current_env()
)

\method{polish_content_pptx}{file_png}(
  x,
  ph = "<p:ph/>",
  pptx,
  ...,
  height = NULL,
  width = NULL,
  units = "in",
  image_fit = c("stretch", "scale"),
  error_call = current_env()
)
}
\arguments{
\item{x}{object to be polished}

\item{ph}{placeholder}

\item{...}{These dots are for future extensions and must be empty.}

\item{full_width}{Logical, used to fit full size of given width}

\item{page_size}{Number of rows to accommodate in a slide}

\item{table_style}{Named list with two entries: style and options. style is
the name of table style to apply as a character, options a character vector
of options to apply from the table style, including: "Banded Columns",
"Banded Rows", "First Column", "Header Row", "Last Column", and "Total Row"}

\item{pptx}{The Powerpoint Presentation object}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{width, height}{defaults to match the height and width of the placeholder
provided. When that cannot be derived, it assumes the plot should be 5
inches tall and 6 inches wide. Users may set the width and height, but they
most both be set.}

\item{res}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{device}{what graphic device to use when saving the plot. Can be one of
"png","svg", or "jpeg".}

\item{scale}{Multiplicative scaling factor to use when saving the plot. See \link[ggplot2:ggsave]{ggplot2::ggsave}.}

\item{escape}{Should \code{x} be escaped with \code{\link[htmltools:htmlEscape]{htmltools::htmlEscape()}}}

\item{collapse}{If \code{TRUE} a \verb{<a:br w:type="textWrapping" w:clear="all">} element is added between the \verb{<a:r>} nodes}

\item{font_color}{color of text. Enter as a hex code.}

\item{font_style}{styling of text, vector including "bold","italic","underline".}

\item{font_size}{point size of text if different from document default.}

\item{font_typeface}{typeface of the text to be added if different from document default.}

\item{list_type}{How to print the list. Options are "none", "unordered" and
"ordered".
\itemize{
\item "none" (default) puts contents on new lines and indents lists accordingly.
\item "unordered" makes the content bulleted list with indents.
\item "ordered" makes a numbered list with intents.
}}

\item{units}{see \link[officer:external_img]{officer::external_img}}

\item{image_fit}{Should the image be distorted to match the dimensions of the placeholder, or scaled up/down and keep dimension ratio (scale). Default is "stretch".}
}
\value{
a "polish_xml_nodeset" of viable PowerPoint OOXML for addition
to PowerPoint presentations.
}
\description{
Polish content for Powerpoint
}
