% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_xml.R
\name{show_xml}
\alias{show_xml}
\title{Show the xml of the polished version of a display in the browser}
\usage{
show_xml(xml, wrap = TRUE)
}
\arguments{
\item{xml}{an xml node or nodeset}

\item{wrap}{if TRUE, the xml is wrapped in a w:body node}
}
\value{
used for side effect of previewing xml in a browser. Helpful for
inspecting the OOXML generated by this package.
}
\description{
Show the xml of the polished version of a display in the browser
}
\details{
This opens the xml in the browser, so that you can do View source
and inspect the structure of the xml
}
