% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx.R
\name{sp_shell}
\alias{sp_shell}
\title{Wrap the placeholder into a <p:sp> shell}
\usage{
sp_shell(ph, error_call = current_env())
}
\arguments{
\item{ph}{placeholder}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Wrap the placeholder into a <p:sp> shell
}
